+++ packages/elements/mention/src/components/ResourcedMentionList/index.tsx
@@ -5,8 +5,8 @@ import { MentionDescription, OnMentionEvent } from '../../types';
 import uniqueId from '../../util/id';
 import debug from '../../util/logger';
 import MentionList from '../MentionList';
-import MentionSpotlight from '../MentionSpotlight';
-import MentionSpotlightController from '../MentionSpotlight/MentionSpotlightController';
+import TeamMentionHighlight from '../TeamMentionHighlight';
+import TeamMentionHighlightController from '../TeamMentionHighlight/TeamMentionHighlightController';
 
 function applyPresence(mentions: MentionDescription[], presences: PresenceMap) {
   const updatedMentions: MentionDescription[] = [];
@@ -42,6 +42,7 @@ export interface Props {
   onSelection?: OnMentionEvent;
   resourceError?: Error;
   isTeamMentionHighlightEnabled?: boolean;
+  createTeamPath?: string; // link to create a team, with context for in-product directories
 }
 
 export interface State {
@@ -223,13 +224,13 @@ export default class ResourcedMentionList extends React.PureComponent<
   };
 
   private closeHighlight = () => {
-    MentionSpotlightController.registerClosed();
+    TeamMentionHighlightController.registerClosed();
   };
 
   private mentionsHighlight = () => {
     const { mentions } = this.state;
-    const { isTeamMentionHighlightEnabled } = this.props;
-    const enabledViaLocalStorage = MentionSpotlightController.isSpotlightEnabled();
+    const { isTeamMentionHighlightEnabled, createTeamPath } = this.props;
+    const enabledViaLocalStorage = TeamMentionHighlightController.isHighlightEnabled();
 
     const shouldShow =
       enabledViaLocalStorage &&
@@ -241,8 +242,8 @@ export default class ResourcedMentionList extends React.PureComponent<
     }
 
     return (
-      <MentionSpotlight
-        createTeamLink="/people/search#createTeam"
+      <TeamMentionHighlight
+        createTeamLink={createTeamPath}
         onClose={this.closeHighlight}
       />
     );