+++ packages/elements/util-data-test/src/mention/story-data.ts
@@ -136,8 +136,8 @@ export const resourceProviderWithInfoHints = new MentionResourceWithInfoHints({
   maxWait: 25,
 });
 
-export const resourceProviderWithSpotlight = new MentionResource({
+export const resourceProviderWithTeamMentionHighlight = new MentionResource({
   minWait: 10,
   maxWait: 25,
-  enableSpotlight: true,
+  enableTeamMentionHighlight: true,
 });